Refactor all objects are loaded from xml file

+++ src/com/spantons/object/ObjectAttributeGetHigh.java
@@ -11,12 +11,12 @@ import com.spantons.singleton.SoundCache;
 
 public class ObjectAttributeGetHigh implements IObjectAttribute {
 
-	private int moveSpeed;
+	private double moveSpeed;
 	private Timer timer;
 	private Object object;
 
 	/****************************************************************************************/
-	public ObjectAttributeGetHigh(int _moveSpeed, Object _object) {
+	public ObjectAttributeGetHigh(double _moveSpeed, Object _object) {
 		moveSpeed = _moveSpeed;
 		object = _object;
 	}
@@ -25,7 +25,7 @@ public class ObjectAttributeGetHigh implements IObjectAttribute {
 	@Override
 	public void loadAttribute(final Entity _entity) {
 		SoundCache.getInstance().getSound(SoundPath.SFX_SMOKE).play();
-		_entity.setMoveSpeed(_entity.getMoveSpeed() - moveSpeed);
+		_entity.setMoveSpeed((int) (_entity.getMoveSpeed() - moveSpeed));
 
 		timer = new Timer(object.timeToConsumable, new ActionListener() {
 
@@ -41,7 +41,7 @@ public class ObjectAttributeGetHigh implements IObjectAttribute {
 	/****************************************************************************************/
 	@Override
 	public void unloadAttribute(Entity _entity) {
-		_entity.setMoveSpeed(_entity.getMoveSpeed() + moveSpeed);
+		_entity.setMoveSpeed((int) (_entity.getMoveSpeed() + moveSpeed));
 	}
 
 }
